🔨Refactor: Service 구현 클래스 분리 및 quey와 command 분리 구현
